Server vakantie update

+++ CIMS_Server 2.0/src/Field_Operations/DAO/TaskDAOImpl.java
@@ -30,7 +30,7 @@ public class TaskDAOImpl implements TaskDAO {
         Query q = em.createNamedQuery("Task.count", Task.class);
         Integer amount = (Integer) q.getSingleResult();
         
-        return amount.intValue();
+        return amount;
     }
 
     @Override
@@ -81,6 +81,5 @@ public class TaskDAOImpl implements TaskDAO {
         List<Task> tasks = q.getResultList();
         
         return tasks;
-    }
-    
+    }    
 }
